# ticketing/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the TicketPulse API for ticketing microservice. Concerts, sessions, sections and tickets are all here.",
        "title": "TicketPulse Ticketing API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/concerts": {
            "get": {
                "description": "Retrieves a list of all concerts.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get all Concerts",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new concert.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a Concert",
                "parameters": [
                    {
                        "description": "Concert Title",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Artist Name",
                        "name": "artist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Image URL",
                        "name": "imageUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/concerts/artist/{artistName}": {
            "get": {
                "description": "Retrieves concerts by artist.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Concerts by Artist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Artist Name",
                        "name": "artistName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/concerts/{id}": {
            "get": {
                "description": "Retrieves details of a concert by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get a Concert by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a concert by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a Concert by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Concert Title",
                        "name": "title",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Artist Name",
                        "name": "artist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Image URL",
                        "name": "imageUrl",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a concert by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a Concert by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/concerts/{id}/sessions": {
            "get": {
                "description": "Retrieves sessions of a concert by its ID.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get Sessions of a Concert",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Concert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ConcertWithSessionListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/ticketing_internal_common_types.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ticketing_internal_common_types.Concert": {
            "type": "object",
            "properties": {
                "artist": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageUrl": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "ticketing_internal_common_types.ConcertListResponse": {
            "type": "object",
            "properties": {
                "concerts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticketing_internal_common_types.Concert"
                    }
                }
            }
        },
        "ticketing_internal_common_types.ConcertResponse": {
            "type": "object",
            "properties": {
                "concert": {
                    "$ref": "#/definitions/ticketing_internal_common_types.Concert"
                }
            }
        },
        "ticketing_internal_common_types.ConcertWithSessionListResponse": {
            "type": "object",
            "properties": {
                "concert": {
                    "type": "string"
                },
                "sessions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ticketing_internal_common_types.Session"
                    }
                }
            }
        },
        "ticketing_internal_common_types.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "ticketing_internal_common_types.Session": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "sessionDateTime": {
                    "type": "integer"
                }
            }
        },
...
        "ticketing_internal_common_types.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}